Adds App Name Animation
