+++ DailyWeather/DailyWeather/Constants.swift
@@ -7,4 +7,5 @@
 
 struct K {
     static let weatherApiKey = "64aca4e110e5a524cc13faaaf78d909f"
+    static let defaultCity = "Mumbai"
 }
