+++ DailyWeather/DailyWeather/View/Base.lproj/Main.storyboard
@@ -32,14 +32,14 @@
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
                             <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Enter City Name" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="RVD-da-6yM">
-                                <rect key="frame" x="80.5" y="356.5" width="253.5" height="46.5"/>
-                                <fontDescription key="fontDescription" name="AvenirNext-Regular" family="Avenir Next" pointSize="34"/>
+                                <rect key="frame" x="110" y="367" width="194" height="36"/>
+                                <fontDescription key="fontDescription" name="AvenirNext-Regular" family="Avenir Next" pointSize="26"/>
                                 <nil key="textColor"/>
                                 <nil key="highlightedColor"/>
                             </label>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="i54-wf-IaA">
-                                <rect key="frame" x="40" y="138" width="334" height="74"/>
-                                <fontDescription key="fontDescription" name="AvenirNext-Medium" family="Avenir Next" pointSize="54"/>
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="i54-wf-IaA">
+                                <rect key="frame" x="20" y="138" width="374" height="0.0"/>
+                                <fontDescription key="fontDescription" name="AvenirNext-Medium" family="Avenir Next" pointSize="48"/>
                                 <nil key="textColor"/>
                                 <nil key="highlightedColor"/>
                             </label>
@@ -69,11 +69,11 @@
                         <viewLayoutGuide key="safeArea" id="M3a-Lb-rQu"/>
                         <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                         <constraints>
-                            <constraint firstItem="M3a-Lb-rQu" firstAttribute="trailing" secondItem="i54-wf-IaA" secondAttribute="trailing" constant="40" id="2Ht-3Q-c55"/>
+                            <constraint firstItem="M3a-Lb-rQu" firstAttribute="trailing" secondItem="i54-wf-IaA" secondAttribute="trailing" constant="20" id="2Ht-3Q-c55"/>
                             <constraint firstItem="M02-Eq-8WF" firstAttribute="centerY" secondItem="DSH-EE-smk" secondAttribute="centerY" id="Bxf-jc-aUS"/>
                             <constraint firstItem="M3a-Lb-rQu" firstAttribute="trailing" secondItem="M02-Eq-8WF" secondAttribute="trailing" constant="22" id="MRx-Gz-hvk"/>
                             <constraint firstItem="M02-Eq-8WF" firstAttribute="leading" secondItem="M3a-Lb-rQu" secondAttribute="leading" constant="22" id="Vzl-oZ-6jM"/>
-                            <constraint firstItem="i54-wf-IaA" firstAttribute="leading" secondItem="M3a-Lb-rQu" secondAttribute="leading" constant="40" id="Xee-Ky-DAm"/>
+                            <constraint firstItem="i54-wf-IaA" firstAttribute="leading" secondItem="M3a-Lb-rQu" secondAttribute="leading" constant="20" id="Xee-Ky-DAm"/>
                             <constraint firstItem="RVD-da-6yM" firstAttribute="centerX" secondItem="DSH-EE-smk" secondAttribute="centerX" id="c9k-2x-3di"/>
                             <constraint firstItem="i54-wf-IaA" firstAttribute="top" secondItem="M3a-Lb-rQu" secondAttribute="top" constant="94" id="eMf-nX-bf6"/>
                             <constraint firstItem="M02-Eq-8WF" firstAttribute="top" secondItem="RVD-da-6yM" secondAttribute="bottom" constant="24" id="hi4-17-Etu"/>
